requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==1.31.2
cytoolz==0.8.2
defusedxml==0.5.0
dill==0.2.7.1
Django==1.11.7
django-allauth==0.34.0
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
djangorestframework==3.7.3
en-core-web-sm==2.0.0
ftfy==4.4.3
fuzzywuzzy==0.15.1
html5lib==0.999999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
numpy==1.13.3
oauthlib==2.0.6
openapi-codec==1.3.2
pathlib==1.0.1
pkg-resources==0.0.0
plac==0.9.6
preshed==1.0.0
psycopg2==2.7.3.2
python-Levenshtein==0.12.0
python3-openid==3.1.0
pytz==2017.3
regex==2017.4.5
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.0
simplejson==3.12.0
six==1.11.0
sklearn==0.0
spacy==2.0.3
termcolor==1.1.0
thinc==6.10.1
toolz==0.8.2
tqdm==4.19.4
ujson==1.35
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
